Safari-611.1.10/LayoutTests/accessibility/ios-simulator/ios-search-predicate-link-children-expected.txt

start text

heading text

sub text
start text 2

link2 text link3
This tests the ability to search for exposed link children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS resultElement.description is 'AXLabel: heading text'
PASS resultElement.description is 'AXLabel: sub text'
PASS resultElement.description is 'AXLabel: heading text'
PASS resultElement.description is 'AXLabel: link2 text'
PASS resultElement.description is 'AXLabel: link3'
PASS successfullyParsed is true

TEST COMPLETE
